Fix int64 error and ran gofmt

+++ main.go
@@ -14,7 +14,7 @@ import (
 
 func main() {
 	f := flags.Get()
-	gofakeit.Seed(f.Seed)
+	gofakeit.Seed(int64(f.Seed))
 	driver := drivers.New(f.Driver)
 	db := connector.Connection(driver, f)
 	defer db.Close()
